pyproject.toml
[project]
name = "ot-p-skai-ontovia-api"
version = "0.1.0"
description = "AgensGraph Agentic AI API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # --- Core Server & Utils ---
    "fastapi==0.115.12",
    "uvicorn==0.34.2",
    "python-dotenv==1.1.0",
    "httpx==0.28.1",

    # --- AI & Agent Framework ---
    "langgraph==0.4.8",
    "langchain-core==0.3.60",
    "langchain-openai==0.3.17",
    "langchain-community==0.3.24",

    # --- Database Drivers ---
    "psycopg2-binary==2.9.10",
    "sqlalchemy==2.0.41",

    # --- Data Validation (Pydantic V2) ---
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["backend"]